#82 merge Domain class with base Symbol class
