--- docs/source/expression_tree/domain.rst
@@ -1,6 +1,0 @@
-Domain
-======
-
-.. autoclass:: pybamm.Domain
-  :members:
-
