+++ pybamm/expression_tree/exceptions.py
@@ -1,0 +1,11 @@
+#
+# Exception classes
+#
+from __future__ import absolute_import, division
+from __future__ import print_function, unicode_literals
+
+
+class DomainError(RuntimeError):
+    """Domain error: an operation was attempted on nodes with un-matched domains"""
+
+    pass
