+++ pybamm/expression_tree/independent_variable.py
@@ -6,13 +6,10 @@ from __future__ import print_function, unicode_literals
 import pybamm
 
 
-class IndependentVariable(pybamm.Domain, pybamm.Symbol):
+class IndependentVariable(pybamm.Symbol):
     """A node in the expression tree represending an independent variable
 
     Used for expressing functions depending on a spatial variable or time
-
-    A variable has a list of domains (text) that it is valid over
-    (inherits from :class:`.Domain`)
 
     Arguments:
 
@@ -22,7 +19,7 @@ class IndependentVariable(pybamm.Domain, pybamm.Symbol):
         list of domains that this variable is valid over
 
 
-    *Extends:* :class:`Domain` and :class:`Symbol`
+    *Extends:* :class:`Symbol`
     """
 
     def __init__(self, name, domain=[]):
